improves readability: splits widgets code in multiple files (one per widget), renames files and references, accordingly
