+++ css/svgsldwidget.css
@@ -1,8 +1,3 @@
-.svg-viewer-widget {
-  border: 2px solid lightgrey;
-  display: inline-block;
-}
-
 .svg-sld-viewer-widget {
   border: 2px solid lightgrey;
   display: inline-block;
