+++ css/svgwidget.css
@@ -1,0 +1,4 @@
+.svg-viewer-widget {
+  border: 2px solid lightgrey;
+  display: inline-block;
+}
