+++ pypowsybl_jupyter/svgsldwidget.py
