+++ pypowsybl_jupyter/svgwidget.py
