+++ src/__tests__/index.spec.ts
@@ -8,15 +8,27 @@
 
 import { createTestModel } from './utils';
 
-import { SvgModel } from '..';
+import { SvgModel, SvgSldModel } from '..';
 
 describe('SvgModel', () => {
   describe('SvgModel', () => {
     it('should be createable with a value', () => {
-      const state = { value: 'Foo Bar!' };
+      const state = { value: 'svg' };
       const model = createTestModel(SvgModel, state);
       expect(model).toBeInstanceOf(SvgModel);
-      expect(model.get('value')).toEqual('Foo Bar!');
+      expect(model.get('value')).toEqual('svg');
+    });
+  });
+});
+
+describe('SvgSldModel', () => {
+  describe('SvgSldModel', () => {
+    it('should be createable with a value', () => {
+      const state = { value: 'svg', value_meta: 'svg_meta' };
+      const model = createTestModel(SvgSldModel, state);
+      expect(model).toBeInstanceOf(SvgSldModel);
+      expect(model.get('value')).toEqual('svg');
+      expect(model.get('value_meta')).toEqual('svg_meta');
     });
   });
 });
